+++ libs/file_permissions/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "kanidm_lib_file_permissions"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[target.'cfg(target_family = "windows")'.dependencies]
+whoami.workspace = true
+
+[target.'cfg(not(target_family = "windows"))'.dependencies]
+users.workspace = true
