+++ server/daemon/Cargo.toml
@@ -18,11 +18,9 @@ name = "kanidmd"
 path = "src/main.rs"
 
 [dependencies]
-# Remove these!
 kanidm_proto.workspace = true
-kanidmd_lib.workspace = true
-
 kanidmd_core.workspace = true
+kanidm_lib_file_permissions.workspace = true
 sketching.workspace = true
 
 clap = { workspace = true, features = ["env"] }
