+++ server/lib/Cargo.toml
@@ -32,10 +32,10 @@ filetime.workspace = true
 futures.workspace = true
 futures-util.workspace = true
 hashbrown.workspace = true
-hex.workspace = true
 idlset.workspace = true
 kanidm_proto.workspace = true
 kanidmd_lib_macros.workspace = true
+kanidm_lib_crypto.workspace = true
 lazy_static.workspace = true
 ldap3_proto.workspace = true
 libc.workspace = true
