+++ server/lib/src/plugins/session.rs
@@ -153,7 +153,7 @@ mod tests {
     use time::OffsetDateTime;
     use uuid::uuid;
 
-    use crate::credential::policy::CryptoPolicy;
+    use kanidm_lib_crypto::CryptoPolicy;
     use crate::credential::Credential;
 
     // Test expiry of old sessions
