+++ tools/cli/Cargo.toml
