+++ tools/cli/Dockerfile
