+++ tools/cli/build.rs
