+++ tools/cli/example.modify.idm_admin.json
