+++ tools/cli/src/cli/badlist.rs
