+++ tools/cli/src/cli/common.rs
