+++ tools/cli/src/cli/domain.rs
