+++ tools/cli/src/cli/lib.rs
