+++ tools/cli/src/cli/raw.rs
