+++ tools/cli/src/cli/recycle.rs
