+++ tools/cli/src/cli/session.rs
