+++ tools/cli/src/opt/kanidm.rs
