+++ tools/cli/src/opt/ssh_authorizedkeys.rs
