+++ tools/cli/src/ssh_authorizedkeys.rs
