+++ unix_integration/src/db.rs
@@ -3,9 +3,9 @@ use std::fmt;
 use std::time::Duration;
 
 use kanidm_proto::v1::{UnixGroupToken, UnixUserToken};
-use kanidmd_lib::be::dbvalue::DbPasswordV1;
-use kanidmd_lib::credential::policy::CryptoPolicy;
-use kanidmd_lib::credential::Password;
+use kanidm_lib_crypto::DbPasswordV1;
+use kanidm_lib_crypto::CryptoPolicy;
+use kanidm_lib_crypto::Password;
 use libc::umask;
 use r2d2::Pool;
 use r2d2_sqlite::SqliteConnectionManager;
